backend/users/pyproject.toml
[project]
name = "users"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Tinn-K", email = "tinn.kanjananuwat@gmail.com"},
]
dependencies = [
    "pydantic>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "fastapi>=0.110.2",
    "uvicorn[standard]>=0.29.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.29",
    "asyncpg>=0.29.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
